+++ backend/src/channel/channel.resolver.ts
@@ -1,4 +1,11 @@
-import { Resolver, Query, Mutation, Args } from '@nestjs/graphql'
+import {
+  Resolver,
+  Query,
+  Mutation,
+  Args,
+  Subscription,
+  Context
+} from '@nestjs/graphql'
 import { ChannelService } from './channel.service'
 import { Channel } from './entities/channel.entity'
 import { CreateChannelInput } from './dto/create-channel.input'
@@ -9,12 +16,58 @@ import {
 } from './dto/update-channel.input'
 import { NanoidValidationPipe } from '../common/pipes/nanoid.pipe'
 import { StringValidationPipe } from '../common/pipes/string.pipe'
-import { AuthorizationGuard } from '../auth/guards/authorization.guard'
+import {
+  AuthorizationGuard,
+  Unprotected
+} from '../auth/guards/authorization.guard'
+import { NanoidsValidationPipe } from 'src/common/pipes/nanoids.pipe'
+import { UsernameValidationPipe } from 'src/common/pipes/username.pipe'
+import { PubSub } from 'graphql-subscriptions'
+import { PrismaService } from 'src/prisma/prisma.service'
+import {
+  ForbiddenAccessChannelAdmin,
+  ForbiddenAccessChannelOwner,
+  ForbiddenAccessData,
+  channelAdminGuard,
+  channelOwnerGuard,
+  userContextGuard
+} from 'src/auth/guards/request.guards'
+import * as bcrypt from 'bcrypt'
+import { EChannelType } from '@prisma/client'
 
 @Resolver(() => Channel)
 @UseGuards(AuthorizationGuard)
 export class ChannelResolver {
-  constructor(private readonly channelService: ChannelService) {}
+  constructor(
+    private readonly channelService: ChannelService,
+    private readonly pubSub: PubSub,
+    private prisma: PrismaService
+  ) {}
+
+  //**************************************************//
+  //  SUBSCRIPTION
+  //**************************************************//
+  @Subscription(() => Channel, {
+    resolve: (payload) => (payload?.res !== undefined ? payload.res : null)
+  })
+  @Unprotected()
+  channelEdition(
+    @Args('id', { type: () => String }, NanoidValidationPipe)
+    id: string
+  ) {
+    return this.pubSub.asyncIterator('channelEdited-' + id)
+  }
+
+  @Subscription(() => Channel, {
+    resolve: (payload) => (payload?.res !== undefined ? payload.res : null)
+  })
+  @Unprotected()
+  channelDeletion(
+    @Args('id', { type: () => String }, NanoidValidationPipe)
+    id: string
+  ) {
+    return this.pubSub.asyncIterator('channelDeleted-' + id)
+  }
 
   //**************************************************//
   //  MUTATION
@@ -22,8 +75,14 @@ export class ChannelResolver {
   @Mutation(() => Channel)
   async createChannel(
     @Args('data', { type: () => CreateChannelInput }, ValidationPipe)
-    data: CreateChannelInput
+    data: CreateChannelInput,
+    @Context() ctx: any
   ): Promise<Channel> {
+    if (!userContextGuard(ctx?.req?.user?.id, data.ownerId))
+      throw new ForbiddenAccessData()
+    if (data?.password) {
+      data.password = bcrypt.hashSync(data.password, 10)
+    }
     return this.channelService.create(data)
   }
 
@@ -32,24 +91,76 @@ export class ChannelResolver {
     @Args('id', { type: () => String }, NanoidValidationPipe)
     id: string,
     @Args('data', { type: () => UpdateChannelInput }, ValidationPipe)
-    data: UpdateChannelInput
+    data: UpdateChannelInput,
+    @Context() ctx: any
   ): Promise<Channel> {
-    return this.channelService.update(id, data)
+    if (!(await channelAdminGuard(ctx?.req?.user?.id, id, this.prisma)))
+      throw new ForbiddenAccessChannelAdmin()
+    const channelMembers = await this.prisma.channelMember.findMany({
+      where: { channelId: id }
+    })
+    if (data?.password) {
+      data.password = bcrypt.hashSync(data.password, 10)
+    }
+
+    const res = await this.channelService.update(id, data)
+
+    await Promise.all(
+      channelMembers.map(async (value) => {
+        await this.pubSub.publish('channelEdited-' + value.userId, { res })
+      })
+    )
+
+    return res
   }
 
   @Mutation(() => Channel)
   async updateChannelOwner(
+    @Args('id', { type: () => String }, NanoidValidationPipe)
     id: string,
-    data: UpdateChannelOwnerIdInput
+    @Args('data', { type: () => UpdateChannelOwnerIdInput }, ValidationPipe)
+    data: UpdateChannelOwnerIdInput,
+    @Context() ctx: any
   ): Promise<Channel | null> {
-    return this.channelService.updateOwner(id, data)
+    if (!(await channelOwnerGuard(ctx?.req?.user?.id, id, this.prisma)))
+      throw new ForbiddenAccessChannelOwner()
+
+    const channelMembers = await this.prisma.channelMember.findMany({
+      where: { channelId: id }
+    })
+
+    const res = await this.channelService.updateOwner(id, data)
+
+    await Promise.all(
+      channelMembers.map(async (value) => {
+        await this.pubSub.publish('channelEdited-' + value.userId, { res })
+      })
+    )
+
+    return res
   }
 
   @Mutation(() => Channel)
   async deleteChannel(
-    @Args('id', { type: () => String }, NanoidValidationPipe) id: string
+    @Args('id', { type: () => String }, NanoidValidationPipe) id: string,
+    @Context() ctx: any
   ): Promise<Channel> {
-    return this.channelService.delete(id)
+    if (!(await channelOwnerGuard(ctx?.req?.user?.id, id, this.prisma)))
+      throw new ForbiddenAccessChannelOwner()
+
+    const channelMembers = await this.prisma.channelMember.findMany({
+      where: { channelId: id }
+    })
+
+    const res = await this.channelService.delete(id)
+
+    await Promise.all(
+      channelMembers.map(async (value) => {
+        await this.pubSub.publish('channelDeleted-' + value.userId, { res })
+      })
+    )
+
+    return res
   }
 
   //**************************************************//
@@ -60,6 +171,21 @@ export class ChannelResolver {
     @Args('id', { type: () => String }, NanoidValidationPipe) id: string
   ): Promise<Channel | null> {
     return this.channelService.findOne(id)
+  }
+
+  @Query(() => Channel)
+  findOneChannelByName(
+    @Args('name', { type: () => String }, UsernameValidationPipe) name: string
+  ): Promise<Channel | null> {
+    return this.channelService.findOneByUsername(name)
+  }
+
+  @Query(() => [Channel])
+  findChannelByChannelIds(
+    @Args('channelIds', { type: () => [String] }, NanoidsValidationPipe)
+    channelIds: string[]
+  ): Promise<Channel[]> {
+    return this.channelService.findChannelByChannelIds(channelIds)
   }
 
   @Query(() => [Channel])
@@ -82,5 +208,13 @@ export class ChannelResolver {
     @Args('userId', { type: () => String }, NanoidValidationPipe) userId: string
   ): Promise<Channel[]> {
     return this.channelService.findAllChannelOfOwner(userId)
+  }
+
+  @Query(() => Boolean)
+  isChannelPasswordSet(
+    @Args('channelId', { type: () => String }, NanoidValidationPipe)
+    channelId: string
+  ): Promise<boolean> {
+    return this.channelService.isChannelPasswordSet(channelId)
   }
 }
