+++ backend/src/private-message/private-message.resolver.ts
@@ -1,4 +1,11 @@
-import { Resolver, Query, Mutation, Args } from '@nestjs/graphql'
+import {
+  Resolver,
+  Query,
+  Mutation,
+  Args,
+  Subscription,
+  Context
+} from '@nestjs/graphql'
 import { PrivateMessageService } from './private-message.service'
 import { PrivateMessage } from './entities/private-message.entity'
 import { CreatePrivateMessageInput } from './dto/create-private-message.input'
@@ -6,12 +13,78 @@ import { UseGuards, ValidationPipe } from '@nestjs/common'
 import { UpdatePrivateMessageInput } from './dto/update-private-message.input'
 import { NanoidValidationPipe } from '../common/pipes/nanoid.pipe'
 import { StringValidationPipe } from '../common/pipes/string.pipe'
-import { AuthorizationGuard } from '../auth/guards/authorization.guard'
+import {
+  AuthorizationGuard,
+  Unprotected
+} from '../auth/guards/authorization.guard'
+import { PubSub } from 'graphql-subscriptions'
+import { ExceptionPrivateMessageDoesNotExist } from '../channel/exceptions/private-message.exception'
+import {
+  ForbiddenAccessData,
+  userContextGuard
+} from 'src/auth/guards/request.guards'
 
 @Resolver(() => PrivateMessage)
 @UseGuards(AuthorizationGuard)
 export class PrivateMessageResolver {
-  constructor(private readonly privateMessageService: PrivateMessageService) {}
+  constructor(
+    private readonly privateMessageService: PrivateMessageService,
+    private readonly pubSub: PubSub
+  ) {}
+
+  //**************************************************//
+  //  SUBSCRIPTION
+  //**************************************************//
+  @Subscription(() => PrivateMessage, {
+    resolve: (payload) => (payload?.res !== undefined ? payload.res : null)
+  })
+  @Unprotected()
+  privateMessageCreation(
+    @Args('senderId', { type: () => String }, NanoidValidationPipe)
+    senderId: string,
+    @Args('receiverId', { type: () => String }, NanoidValidationPipe)
+    receiverId: string
+  ) {
+    const [notificationSenderId, notificationReceiverId] =
+      senderId > receiverId ? [receiverId, senderId] : [senderId, receiverId]
+    return this.pubSub.asyncIterator(
+      'messageReceived-' + notificationSenderId + notificationReceiverId
+    )
+  }
+
+  @Subscription(() => PrivateMessage, {
+    resolve: (payload) => (payload?.res !== undefined ? payload.res : null)
+  })
+  @Unprotected()
+  privateMessageEdition(
+    @Args('senderId', { type: () => String }, NanoidValidationPipe)
+    senderId: string,
+    @Args('receiverId', { type: () => String }, NanoidValidationPipe)
+    receiverId: string
+  ) {
+    const [notificationSenderId, notificationReceiverId] =
+      senderId > receiverId ? [receiverId, senderId] : [senderId, receiverId]
+    return this.pubSub.asyncIterator(
+      'messageEdited-' + notificationSenderId + notificationReceiverId
+    )
+  }
+
+  @Subscription(() => PrivateMessage, {
+    resolve: (payload) => (payload?.res !== undefined ? payload.res : null)
+  })
+  @Unprotected()
+  privateMessageDeletion(
+    @Args('senderId', { type: () => String }, NanoidValidationPipe)
+    senderId: string,
+    @Args('receiverId', { type: () => String }, NanoidValidationPipe)
+    receiverId: string
+  ) {
+    const [notificationSenderId, notificationReceiverId] =
+      senderId > receiverId ? [receiverId, senderId] : [senderId, receiverId]
+    return this.pubSub.asyncIterator(
+      'messageDeleted-' + notificationSenderId + notificationReceiverId
+    )
+  }
 
   //**************************************************//
   //  MUTATION
@@ -19,9 +92,24 @@ export class PrivateMessageResolver {
   @Mutation(() => PrivateMessage)
   async createPrivateMessage(
     @Args('data', { type: () => CreatePrivateMessageInput }, ValidationPipe)
-    data: CreatePrivateMessageInput
+    data: CreatePrivateMessageInput,
+    @Context() ctx: any
   ): Promise<PrivateMessage> {
-    return this.privateMessageService.create(data)
+    if (!userContextGuard(ctx?.req?.user?.id, data.senderId))
+      throw new ForbiddenAccessData()
+
+    const { senderId, receiverId } = data
+    const [notificationSenderId, notificationReceiverId] =
+      senderId > receiverId ? [receiverId, senderId] : [senderId, receiverId]
+
+    const res = await this.privateMessageService.create(data)
+
+    await this.pubSub.publish(
+      'messageReceived-' + notificationSenderId + notificationReceiverId,
+      { res }
+    )
+
+    return res
   }
 
   @Mutation(() => PrivateMessage)
@@ -29,16 +117,50 @@ export class PrivateMessageResolver {
     @Args('id', { type: () => String }, NanoidValidationPipe)
     id: string,
     @Args('data', { type: () => UpdatePrivateMessageInput }, ValidationPipe)
-    data: UpdatePrivateMessageInput
+    data: UpdatePrivateMessageInput,
+    @Context() ctx: any
   ): Promise<PrivateMessage> {
-    return this.privateMessageService.update(id, data)
+    const msg = await this.findOnePrivateMessage(id)
+    if (!msg) throw new ExceptionPrivateMessageDoesNotExist()
+
+    if (!userContextGuard(ctx?.req?.user?.id, msg.senderId))
+      throw new ForbiddenAccessData()
+
+    const { senderId, receiverId } = msg
+    const [notificationSenderId, notificationReceiverId] =
+      senderId > receiverId ? [receiverId, senderId] : [senderId, receiverId]
+
+    const res = await this.privateMessageService.update(id, data)
+
+    await this.pubSub.publish(
+      'messageEdited-' + notificationSenderId + notificationReceiverId,
+      { res }
+    )
+    return res
   }
 
   @Mutation(() => PrivateMessage)
   async deletePrivateMessage(
-    @Args('id', { type: () => String }, NanoidValidationPipe) id: string
+    @Args('id', { type: () => String }, NanoidValidationPipe) id: string,
+    @Context() ctx: any
   ): Promise<PrivateMessage> {
-    return this.privateMessageService.delete(id)
+    const msg = await this.findOnePrivateMessage(id)
+    if (!msg) throw new ExceptionPrivateMessageDoesNotExist()
+
+    if (!userContextGuard(ctx?.req?.user?.id, msg.senderId))
+      throw new ForbiddenAccessData()
+
+    const { senderId, receiverId } = msg
+    const [notificationSenderId, notificationReceiverId] =
+      senderId > receiverId ? [receiverId, senderId] : [senderId, receiverId]
+
+    const res = await this.privateMessageService.delete(id)
+
+    await this.pubSub.publish(
+      'messageDeleted-' + notificationSenderId + notificationReceiverId,
+      { res }
+    )
+    return res
   }
 
   //**************************************************//
