+++ backend/src/relation-blocked/relation-blocked.resolver.ts
@@ -1,17 +1,46 @@
-import { Resolver, Query, Mutation, Args } from '@nestjs/graphql'
+import {
+  Resolver,
+  Query,
+  Mutation,
+  Args,
+  Subscription,
+  Context
+} from '@nestjs/graphql'
 import { RelationBlockedService } from './relation-blocked.service'
 import { RelationBlocked } from './entities/relation-blocked.entity'
 import { RelationBlockedInput } from './dto/create-relation-blocked.input'
 import { UseGuards, ValidationPipe } from '@nestjs/common'
 import { NanoidValidationPipe } from '../common/pipes/nanoid.pipe'
-import { AuthorizationGuard } from '../auth/guards/authorization.guard'
+import {
+  AuthorizationGuard,
+  Unprotected
+} from '../auth/guards/authorization.guard'
+import { PubSub } from 'graphql-subscriptions'
+import {
+  ForbiddenAccessData,
+  userContextGuard
+} from 'src/auth/guards/request.guards'
 
 @Resolver(() => RelationBlocked)
 @UseGuards(AuthorizationGuard)
 export class RelationBlockedResolver {
   constructor(
-    private readonly relationBlockedService: RelationBlockedService
+    private readonly relationBlockedService: RelationBlockedService,
+    private readonly pubSub: PubSub
   ) {}
+
+  //**************************************************//
+  //  SUBSCRIPTION
+  //**************************************************//
+  @Subscription(() => RelationBlocked, {
+    resolve: (payload) => (payload?.res !== undefined ? payload.res : null)
+  })
+  @Unprotected()
+  relationBlockedCreation(
+    @Args('userId', { type: () => String }, NanoidValidationPipe) userId: string
+  ) {
+    return this.pubSub.asyncIterator('blockedReceived-' + userId)
+  }
 
   //**************************************************//
   //  MUTATION
@@ -19,9 +48,21 @@ export class RelationBlockedResolver {
   @Mutation(() => RelationBlocked)
   async createRelationBlocked(
     @Args('data', { type: () => RelationBlockedInput }, ValidationPipe)
-    data: RelationBlockedInput
+    data: RelationBlockedInput,
+    @Context() ctx: any
   ): Promise<RelationBlocked> {
-    return this.relationBlockedService.create(data)
+    if (!userContextGuard(ctx?.req?.user?.id, data.userBlockingId))
+      throw new ForbiddenAccessData()
+
+    const res = await this.relationBlockedService.create(data)
+
+    await this.pubSub.publish('blockedReceived-' + data.userBlockedId, {
+      res
+    })
+    await this.pubSub.publish('blockedReceived-' + data.userBlockingId, {
+      res
+    })
+    return res
   }
 
   @Mutation(() => RelationBlocked)
@@ -29,9 +70,14 @@ export class RelationBlockedResolver {
     @Args('userAId', { type: () => String }, NanoidValidationPipe)
     userAId: string,
     @Args('userBId', { type: () => String }, NanoidValidationPipe)
-    userBId: string
+    userBId: string,
+    @Context() ctx: any
   ): Promise<RelationBlocked> {
-    return this.relationBlockedService.delete(userAId, userBId)
+    if (!userContextGuard(ctx?.req?.user?.id, userAId))
+      throw new ForbiddenAccessData()
+
+    const res = await this.relationBlockedService.delete(userAId, userBId)
+    return res
   }
 
   //**************************************************//
